+++ src/Stub/Factory/StubFactoryInterface.php
@@ -21,10 +21,15 @@ interface StubFactoryInterface
     /**
      * Create a new stub.
      *
-     * @param callable|null $callback The callback, or null to create an unbound stub.
-     * @param mixed         $self     The self value.
+     * @param callable|null $callback              The callback, or null to create an unbound stub.
+     * @param mixed         $self                  The self value.
+     * @param callable|null $defaultAnswerCallback The callback to use when creating a default answer.
      *
      * @return StubInterface The newly created stub.
      */
-    public function create($callback = null, $self = null);
+    public function create(
+        $callback = null,
+        $self = null,
+        $defaultAnswerCallback = null
+    );
 }